Corrections throughout to types used Validate the length in the UDP datagram header as being greater than or equal to minimum Streamline processing of additional bytes in headers of IPv4 and TCP packets Minor tweaks to formatting for switch statements

+++ PacketCaptureAnalyser/Ethernet Frame/IPv4 Packet/ICMPv4 Packet/ICMPv4PacketProcessing.cs
@@ -34,7 +34,7 @@ namespace EthernetFrameNamespace.IPv4PacketNamespace.ICMPv4PacketNamespace
             this.TheBinaryReader = TheBinaryReader;
         }
 
-        public bool Process(int TheICMPv4PacketLength)
+        public bool Process(ushort TheICMPv4PacketLength)
         {
             bool TheResult = true;
